renamed rule as ContainsShouldRule, fixed tests
