+++ src/test/java/opinion/rules/RuleFactoryTest.java
